requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
pydantic-settings>=2.0.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
alembic>=1.7.0,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-multipart>=0.0.5,<0.0.6
email-validator>=1.1.3,<1.2.0
requests>=2.26.0,<2.27.0
pytest>=6.2.5,<6.3.0
httpx>=0.18.2,<0.19.0
slowapi>=0.1.4,<0.2.0
numpy>=1.21.0,<1.22.0
pandas>=1.3.0,<1.4.0
yfinance>=0.1.70
alpha_vantage>=2.3.1
transformers>=4.15.0
torch>=1.10.0
openai>=0.27.0
python-dotenv>=0.19.0
redis>=4.0.0
plotly>=5.3.0
ta>=0.7.0
scikit-learn>=0.24.0
gunicorn>=20.1.0
whitenoise>=6.0.0 